# ddl: decouple the index writer from range task