<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Problem Summary:

Since this is the refactoring, let's skip the issue check.

### What is changed and how it works?

![image](https://user-images.githubusercontent.com/24713065/227874811-fafbd5f7-b1a7-4eee-a44d-8392d57d8f00.png)

This PR removes the red part in the above image.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

I made a simple performance test on 10M sysbench dataset.

Before this PR:
```
mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 35.43 sec)

mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 12.41 sec)

mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 42.54 sec)

mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 35.92 sec)
```

After this PR:
```
mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 17.92 sec)

mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 37.46 sec)

mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 45.83 sec)

mysql> alter table sbtest1 add index idx(c);
Query OK, 0 rows affected (3 min 16.46 sec)
```

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```